Implementação do tratamento de exceções e soft delete

+++ src/main/kotlin/CRUD/PacientesAPI/Kotlim/Service/MedicoService.kt
@@ -2,52 +2,50 @@ package CRUD.PacientesAPI.Kotlim.Service
 
 import CRUD.PacientesAPI.Kotlim.Domain.DTOs.MedicoDTO
 import CRUD.PacientesAPI.Kotlim.Domain.DTOs.MedicoUpdateDTO
-import CRUD.PacientesAPI.Kotlim.Domain.DTOs.PacienteDTO
 import CRUD.PacientesAPI.Kotlim.Domain.Entity.MedicoEntity
-import CRUD.PacientesAPI.Kotlim.Domain.Mapper.toDTO
+import CRUD.PacientesAPI.Kotlim.Domain.Mapper.toAgendamentoDTO
 import CRUD.PacientesAPI.Kotlim.Domain.Mapper.toEntity
 import CRUD.PacientesAPI.Kotlim.Domain.Mapper.toupDTO
 import CRUD.PacientesAPI.Kotlim.Domain.repository.MedicoRepository
+import CRUD.PacientesAPI.Kotlim.Infra.ConflictException
+import CRUD.PacientesAPI.Kotlim.Infra.ResourceNotFoundException
 import org.springframework.stereotype.Service
 import java.lang.RuntimeException
-import kotlin.jvm.optionals.getOrNull
 
 
 @Service
 class MedicoService(private val repository: MedicoRepository) {
 
-    fun findAll(): List<MedicoDTO> {
-      var findAll =  repository.findAll()
-                     .map{ it.toDTO()}
-        return findAll
-    }
+    fun findAll(): List<MedicoDTO> { return  repository.findAll().map{ it.toAgendamentoDTO()} }
 
     fun findById(id: Long): MedicoDTO? {
         val findbyid = repository.findById(id)
-        val response = findbyid.map {it.toDTO() }.orElse(null)
+        val response = findbyid.map {it.toAgendamentoDTO() }.orElse(null)
         return response
     }
 
     fun salvar(dto: MedicoDTO): MedicoDTO {
         var entity = dto.toEntity()
         if (repository.existsByCrm(dto.crm)){
-            throw RuntimeException("CRM ja cadastrada")
+            throw ConflictException("CRM ja cadastrada")
         }
         if (repository.existsByEmail(dto.email)){
-            throw RuntimeException("email ja cadastrada")
+            throw ConflictException("email ja cadastrada")
         }
         var save = repository.save(entity)
-        return save.toDTO()
+        return save.toAgendamentoDTO()
     }
 
-    fun delete(id: Long){
-      return  repository.deleteById(id)
+    fun delete(id: Long) {
+        if (!repository.existsById(id))
+            throw ResourceNotFoundException("Médico não encontrado com id $id") // novo
+        repository.deleteById(id)
     }
 
     fun alterar(id: Long, dto: MedicoUpdateDTO) : MedicoUpdateDTO {
 
         val medico : MedicoEntity = repository.findById(id)
-            .orElseThrow{RuntimeException("Medico nao encontrado")}
+            .orElseThrow { ResourceNotFoundException("Médico não encontrado com id $id") }
 
         medico.nome = dto.nome ?: medico.nome
         medico.crm = dto.crm ?: medico.crm
@@ -57,10 +55,15 @@ class MedicoService(private val repository: MedicoRepository) {
 
         val saved = repository.save(medico)
         return saved.toupDTO()
+    }
 
+    fun softDelete(id: Long) {
+        val medico : MedicoEntity = repository.findById(id)
+            . orElseThrow { ResourceNotFoundException("Médico nao encontrado com id $id") }
 
+        medico.ativa = false
 
-
+        repository.save(medico)
     }
 }
